Moved AbooFileUploaderService to a separate file since CodeFactor did not allow to have this and the FileUploader in the same class
