+++ ShareX.UploadersLib/ShareX.UploadersLib.csproj
@@ -126,6 +126,7 @@
     <Compile Include="BaseUploaders\GenericUploader.cs" />
     <Compile Include="BaseUploaders\URLSharer.cs" />
     <Compile Include="FileUploaders\Aboo.cs" />
+    <Compile Include="FileUploaders\AbooFileUploaderService.cs" />
     <Compile Include="FileUploaders\AmazonS3.cs" />
     <Compile Include="FileUploaders\AmazonS3Endpoint.cs" />
     <Compile Include="FileUploaders\AmazonS3Settings.cs" />
